fix(discord): gate tmux reaper wiring on unix (#4485)
