+++ src/services/discord/tmux_reaper.rs
@@ -150,84 +150,6 @@ pub(in crate::services::discord) async fn heal_stale_busy_mailbox(
     .await
 }
 
-/// Claim an intake turn, healing a stale busy mailbox and retrying once when
-/// its managed tmux session is proven absent. The heal path retains the
-/// identity and respawn guards in `heal_stale_busy_mailbox`; callers only need
-/// to supply the inputs already available at the intake claim site.
-pub(in crate::services::discord) async fn try_start_turn_with_stale_busy_heal(
-    shared: &Arc<SharedData>,
-    channel_id: serenity::ChannelId,
-    cancel_token: Arc<CancelToken>,
-    request_owner: serenity::UserId,
-    user_msg_id: serenity::MessageId,
-    context: (Option<&str>, &ProviderKind, Option<&str>),
-) -> bool {
-    let (session_key, provider, tmux_session_name) = context;
-    let started = super::router::mailbox_try_start_turn_with_terminal_marker_cleanup(
-        shared,
-        channel_id,
-        cancel_token.clone(),
-        request_owner,
-        user_msg_id,
-        session_key,
-    )
-    .await;
-
-    #[cfg(unix)]
-    if !started
-        && let Some(tmux_session_name) = tmux_session_name
-        && heal_stale_busy_mailbox(
-            shared,
-            provider,
-            channel_id,
-            tmux_session_name,
-            "discord_intake",
-        )
-        .await
-    {
-        return super::router::mailbox_try_start_turn_with_terminal_marker_cleanup(
-            shared,
-            channel_id,
-            cancel_token,
-            request_owner,
-            user_msg_id,
-            session_key,
-        )
-        .await;
-    }
-
-    #[cfg(not(unix))]
-    let _ = (provider, tmux_session_name);
-    started
-}
-
-pub(in crate::services::discord) fn stale_busy_context<'a>(
-    provider: &'a ProviderKind,
-    session_names: [Option<&'a str>; 2],
-) -> (Option<&'a str>, &'a ProviderKind, Option<&'a str>) {
-    (session_names[0], provider, session_names[1])
-}
-
-pub(in crate::services::discord) async fn resolve_channel_tmux_names(
-    shared: &Arc<SharedData>,
-    provider: &ProviderKind,
-    channel_id: serenity::ChannelId,
-) -> (Option<String>, Option<String>) {
-    let data = shared.core.lock().await;
-    let channel_name = data
-        .sessions
-        .get(&channel_id)
-        .and_then(|session| session.channel_name.clone());
-    let tmux_session_name = if provider.uses_managed_tmux_backend() {
-        channel_name
-            .as_ref()
-            .map(|name| provider.build_tmux_session_name(name))
-    } else {
-        None
-    };
-    (channel_name, tmux_session_name)
-}
-
 /// Idle-channel half of #4485: unlike the legacy dead-session pass below,
 /// this starts from active mailboxes, so a fully removed tmux session (absent
 /// from `list-sessions`) is still discoverable and its global concurrency slot
